pyproject.toml
[project]
name = "email-agent"
version = "0.1.0"
description = "一个智能邮件处理工具，通过 IMAP 拉取未读邮件，使用智能聚类算法将相关邮件分组，并利用 LLM 生成优先级待办清单或邮件摘要"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "imapclient>=2.3.1",
    "beautifulsoup4>=4.12.2",
    "python-dateutil>=2.9.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "openai>=1.51.0",
    "python-dotenv>=1.0.1",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = []

